pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
autopep8 = "^1.7.0"
django-extensions = "^3.2.1"
ipython = {extras = ["terminal"], version = "^8.5.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
